data_processing/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
et-xmlfile==1.1.0
idna==3.6
install==1.3.5
joblib==1.3.2
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
pdfminer.six==20231228
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soupsieve==2.5
threadpoolctl==3.2.0
tzdata==2023.4
urllib3==2.1.0
